--- sqrl-base/src/main/java/com/datasqrl/json/NativeType.java
@@ -1,10 +1,0 @@
-package com.datasqrl.json;
-
-public interface NativeType<T> {
-  Class getType();
-
-  // Downcast to a SQL type if a connector cannot natively handle it
-  T unknownTypeDowncast();
-
-  String getPhysicalTypeName(String name);
-}
