+++ sqrl-engines/sqrl-engine-flink/pom.xml
@@ -46,6 +46,11 @@
     </dependency>
 
     <dependency>
+      <groupId>com.datasqrl</groupId>
+      <artifactId>sqrl-vector-execute</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.apache.flink</groupId>
       <artifactId>flink-core</artifactId>
     </dependency>
