+++ sqrl-lib/sqrl-vector/sqrl-vector-planner/src/main/java/com/datasqrl/vector/CenterSqlTranslation.java
@@ -1,5 +1,4 @@
-package com.datasqrl.function.vector;
-
+package com.datasqrl.vector;
 import static com.datasqrl.function.CalciteFunctionUtil.lightweightOp;
 
 import com.datasqrl.function.PgSpecificOperatorTable;
@@ -7,21 +6,20 @@ import com.datasqrl.function.translations.PostgresSqlTranslation;
 import com.datasqrl.function.translations.SqlTranslation;
 import com.google.auto.service.AutoService;
 import org.apache.calcite.sql.SqlCall;
-import org.apache.calcite.sql.SqlLiteral;
 import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.calcite.sql.parser.SqlParserPos;
 
 @AutoService(SqlTranslation.class)
-public class CosineDistanceSqlTranslation extends PostgresSqlTranslation {
+public class CenterSqlTranslation extends PostgresSqlTranslation {
 
-  public CosineDistanceSqlTranslation() {
-    super(lightweightOp("cosinedistance"));
+  public CenterSqlTranslation() {
+    super(lightweightOp("center"));
   }
 
   @Override
   public void unparse(SqlCall call, SqlWriter writer, int leftPrec, int rightPrec) {
-    PgSpecificOperatorTable.CosineDistance.createCall(SqlParserPos.ZERO, call.getOperandList())
+    SqlStdOperatorTable.AVG.createCall(SqlParserPos.ZERO, call.getOperandList())
         .unparse(writer, leftPrec, rightPrec);
   }
 }
